Funcionando Sencilla

+++ src/main/java/com/turismo/operadora/repository/ReservaRepository.java
@@ -2,6 +2,6 @@ package com.turismo.operadora.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.turismo.operadora.entity.Reserva;
-public interface ReservaRepository extends JpaRepository<Reserva, int> {
+public interface ReservaRepository extends JpaRepository<Reserva,Long> {
 
 }
